+++ RELEASE_NOTES.md
@@ -4,9 +4,10 @@
 # MongoDB Enterprise Kubernetes Operator 1.19.0
 
 ## Breaking changes and deprecations
+* Making the field orgID in the project configmap a requirement. **Note**: If explicitly an empty `orgID = ""` has been chosen then OM will try to create an ORG with the project name.
 * Ubuntu-based images were deprecated in favor of UBI-based images in operator version 1.17.0. In the 1.19.0 release we are removing the support for Ubuntu-based images. The ubuntu based images won't be rebuilt daily with updates. Please upgrade to the UBI-based images by following these instructions: https://www.mongodb.com/docs/kubernetes-operator/master/tutorial/migrate-k8s-images/#migrate-k8s-images
 
-<!-- Past releases -->
+<!-- Past Releases -->
 # MongoDB Enterprise Kubernetes Operator 1.18.0
 
 ## Improvements
@@ -28,7 +29,6 @@
 [1.17.0](https://www.mongodb.com/docs/kubernetes-operator/master/release-notes/#k8s-op-full-1-17-0). Before upgrading to this version, make
 sure you migrated to the new TLS format using the following [Migration Guide](https://www.mongodb.com/docs/kubernetes-operator/v1.16/tutorial/migrate-to-new-tls-format/) before upgrading the Operator.
 
-<!-- Past Releases -->
 # MongoDB Enterprise Kubernetes Operator 1.17.2
 
 * Fixed the OpenShift installation problem mentioned in the Enterprise Operator 1.7.1 release notes. The OLM (Operator Lifecycle Manager)